chart now responds to viewport resize
