+++ src/tsomi.css
@@ -104,7 +104,7 @@ body {
 
 #wikidiv {
   /* position: absolute; */
-  width: 35%;
+  width: 25%;
   height: 100%;
 	float: right;
 }
